requirements.txt
openai==1.2.3
python-dotenv==1.0.0
flask==3.0.0
flask_restful==0.3.10
flask_cors==4.0.0